Implement LWG-4135 and improve test coverage for P1209R0
